requirements.txt
google-generativeai==0.3.2
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.0
tiktoken==0.8.0
tree-sitter>=0.23.2
tree-sitter-language-pack==0.9.0
faiss-cpu==1.8.0
sentence-transformers==2.3.1
torch>=1.9.0
transformers>=4.21.0
numpy<2.0.0
scikit-learn>=1.0.0
huggingface_hub>=0.19.0,<1.0.0
watchdog
qdrant-client==1.7.0